fix: api-fixes
- Fix http response codes
- Update Jenkinsfile to prune docker images and caches
- Update Jenkinsfile to run build and push stages on master only

+++ src/api/controllers/httpcheck.controller.js
@@ -4,24 +4,7 @@ import {
   ResourceNotFoundError,
   ServiceUnavailableError,
 } from '../utils/error.util'
-
-/**
- * `getAll` controller for GET `/v1/http-check` endpoint
- * @param {*} req API request parameter
- * @param {*} res API response
- * @param {*} next API errorHandler middleware
- * @returns HTTP 200 OK, 503 Service Unavailable
- */
-export const getAll = async (req, res, next) => {
-  try {
-    const httpChecks = await service.getAll()
-    if (!httpChecks) throw new ServiceUnavailableError()
-
-    return res.status(200).json(httpChecks)
-  } catch (error) {
-    next(error)
-  }
-}
+import { httpStatusCodes } from '../../constants/responseCodes'
 
 /**
  * `get` controller for GET `/v1/http-check/:id` endpoint
@@ -32,10 +15,17 @@ export const getAll = async (req, res, next) => {
  */
 export const get = async (req, res, next) => {
   try {
-    const httpCheck = await service.findById(req.params.id)
-    if (!httpCheck) throw new ResourceNotFoundError(`HttpCheck not found`)
+    if (req.params.id) {
+      const httpCheck = await service.findById(req.params.id)
+      if (!httpCheck) throw new ResourceNotFoundError(`HttpCheck not found`)
 
-    return res.status(200).json(httpCheck)
+      return res.status(200).json(httpCheck)
+    }
+
+    const httpChecks = await service.getAll()
+    if (!httpChecks) throw new ServiceUnavailableError()
+
+    return res.status(200).json(httpChecks)
   } catch (error) {
     next(error)
   }
@@ -50,6 +40,11 @@ export const get = async (req, res, next) => {
  */
 export const create = async (req, res, next) => {
   try {
+    if (!httpStatusCodes.includes(req.body.response_status_code))
+      throw new BadRequestError(
+        'response_status_code must be a valid HTTP response code'
+      )
+
     const httpCheck = await service.create(req.body)
     if (!httpCheck) throw new ServiceUnavailableError()
 
@@ -68,6 +63,13 @@ export const create = async (req, res, next) => {
  */
 export const update = async (req, res, next) => {
   try {
+    if (!req.params.id) throw new BadRequestError('Invalid id')
+
+    if (!httpStatusCodes.includes(req.body.response_status_code))
+      throw new BadRequestError(
+        'response_status_code must be a valid HTTP response code'
+      )
+
     const updatedHttpCheck = await service.update(req.params.id, req.body)
     if (!updatedHttpCheck) throw new ServiceUnavailableError()
 
@@ -77,7 +79,7 @@ export const update = async (req, res, next) => {
   }
 }
 
-/**
+/**contains
  * `remove` controller for DELETE `/v1/http-check/:id` endpoint
  * @param {*} req API request parameter
  * @param {*} res API response
@@ -86,8 +88,11 @@ export const update = async (req, res, next) => {
  */
 export const remove = async (req, res, next) => {
   try {
+    if (!req.params.id) throw new BadRequestError('Invalid id')
+
     const deletedCount = await service.remove(req.params.id)
-    if (!deletedCount) throw new BadRequestError("HttpCheck doesn't exist")
+    if (!deletedCount)
+      throw new ResourceNotFoundError("HttpCheck doesn't exist")
 
     return res.sendStatus(204)
   } catch (error) {
